iam | begin refactoring to repository pattern
