+++ apps/iam/src/resources/permission/repository/entities/permission.entity.ts
@@ -8,7 +8,7 @@ import {
   Unique,
   UpdateDateColumn,
 } from 'typeorm';
-import { User } from '../../user/entities/user.entity';
+import { User } from '../../../user/repository/entities/user.entity';
 
 @Entity()
 @Unique(['resource', 'action'])
