+++ apps/iam/src/resources/user/resolvers/user-delivery-window.resolver.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { randomUUID } from 'crypto';
 import { UserDeliveryWindowLoader } from '../data-loaders/user-delivery-window.loader';
-import { DeliveryWindow } from '../entities/delivery-window.entity';
-import { User } from '../entities/user.entity';
+import { DeliveryWindow } from '../repository/entities/delivery-window.entity';
+import { User } from '../repository/entities/user.entity';
 import { UserDeliveryWindowResolver } from './user-delivery-window.resolver';
 
 type MockUserDeliveryWindowLoader = Partial<
