+++ apps/iam/src/resources/user/user.module.ts
@@ -4,8 +4,8 @@ import { CommonModule } from '../../common/common.module';
 import { PermissionModule } from '../permission/permission.module';
 import { UserDeliveryWindowLoader } from './data-loaders/user-delivery-window.loader';
 import { UserPermissionLoader } from './data-loaders/user-permission.loader';
-import { DeliveryWindow } from './entities/delivery-window.entity';
-import { User } from './entities/user.entity';
+import { DeliveryWindow } from './repository//entities/delivery-window.entity';
+import { User } from './repository/entities/user.entity';
 import { UserDeliveryWindowResolver } from './resolvers/user-delivery-window.resolver';
 import { UserPermissionResolver } from './resolvers/user-permission.resolver';
 import { UserResolver } from './user.resolver';
