+++ apps/iam/src/resources/user/user.resolver.ts
@@ -5,7 +5,7 @@ import { Args, ID, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { GraphQLError } from 'graphql';
 import { CreateUserInput } from './dto/create-user.input';
 import { UpdateUserInput } from './dto/update-user.input';
-import { User } from './entities/user.entity';
+import { User } from './repository/entities/user.entity';
 import { UserService } from './user.service';
 
 @Resolver()
